+++ tests/self-target.nix
@@ -17,10 +17,16 @@ let
       };
   };
 in
-assert
-  builtins.sort builtins.lessThan nodes.application.config.networking.hosts."192.0.2.10" == [
-    "application-hostname.example"
-    "local.example"
-  ];
-assert checkAssertions nodes;
-true
+{
+  testHosts = {
+    expr = builtins.sort builtins.lessThan nodes.application.config.networking.hosts."192.0.2.10";
+    expected = [
+      "application-hostname.example"
+      "local.example"
+    ];
+  };
+  testAssertions = {
+    expr = checkAssertions nodes;
+    expected = true;
+  };
+}